NodeNs_3D/requirements.txt
matplotlib==3.7.2
numpy==1.20.3
paho_mqtt==1.6.1
setuptools

